Move inner classes from FunctionParameterResolver to package
