+++ fhirpath/src/main/java/au/csiro/pathling/fhirpath/function/resolver/FunctionParameterResolver.java
@@ -1,4 +1,4 @@
-package au.csiro.pathling.fhirpath.function;
+package au.csiro.pathling.fhirpath.function.resolver;
 
 import static java.util.Objects.isNull;
 
@@ -10,6 +10,8 @@ import au.csiro.pathling.fhirpath.TypeSpecifier;
 import au.csiro.pathling.fhirpath.collection.BooleanCollection;
 import au.csiro.pathling.fhirpath.collection.Collection;
 import au.csiro.pathling.fhirpath.collection.EmptyCollection;
+import au.csiro.pathling.fhirpath.function.CollectionTransform;
+import au.csiro.pathling.fhirpath.function.FunctionInput;
 import au.csiro.pathling.fhirpath.path.ParserPaths;
 import jakarta.annotation.Nonnull;
 import jakarta.annotation.Nullable;
@@ -39,127 +41,6 @@ public class FunctionParameterResolver {
    */
   List<FhirPath> actualArguments;
 
-
-  /**
-   * Represents a function method with its bound arguments, ready for invocation.
-   */
-  @Value(staticConstructor = "of")
-  public static class FunctionInvocation {
-
-    /**
-     * The method to be invoked
-     */
-    Method method;
-
-    /**
-     * The resolved arguments to pass to the method
-     */
-    Object[] arguments;
-
-    /**
-     * Invokes the function with the bound arguments.
-     *
-     * @return The result of the function invocation as a Collection
-     * @throws RuntimeException if the invocation fails
-     */
-    public Collection invoke() {
-      try {
-        return (Collection) method.invoke(null, arguments);
-      } catch (final IllegalAccessException | InvocationTargetException e) {
-        throw new RuntimeException("Error invoking function: " + method.getName(), e);
-      }
-    }
-  }
-
-  /**
-   * Provides context for parameter binding operations, including error reporting.
-   * <p>
-   * This class helps generate consistent, informative error messages that include the function name
-   * and context (input or specific argument).
-   */
-  @Value
-  public static class BindingContext {
-
-    /**
-     * The method being bound
-     */
-    @Nonnull
-    Method method;
-
-    /**
-     * Description of the current binding context (e.g., "input" or "argument 0")
-     */
-    @Nullable
-    String contextDescription;
-
-    /**
-     * Creates a binding context for a method.
-     *
-     * @param method The method being bound
-     * @return A new binding context
-     */
-    @Nonnull
-    public static BindingContext forMethod(@Nonnull final Method method) {
-      return new BindingContext(method, null);
-    }
-
-    /**
-     * Creates a binding context for the input of a method.
-     *
-     * @return A new binding context for the input
-     */
-    @Nonnull
-    public BindingContext forInput() {
-      return new BindingContext(method, "input");
-    }
-
-    /**
-     * Creates a binding context for an argument of a method.
-     *
-     * @param index The index of the argument
-     * @param parameterType The type of the parameter
-     * @return A new binding context for the argument
-     */
-    @Nonnull
-    public BindingContext forArgument(final int index, @Nonnull final Class<?> parameterType) {
-      return new BindingContext(method,
-          "argument " + index + " (" + parameterType.getSimpleName() + ")");
-    }
-
-    /**
-     * Reports an error in this binding context.
-     *
-     * @param issue The issue to report
-     * @return Never returns, always throws an exception
-     * @throws InvalidUserInputError Always thrown with a formatted message
-     */
-    public <T> T reportError(@Nonnull final String issue) {
-      final StringBuilder message = new StringBuilder("Function '")
-          .append(method.getName())
-          .append("'");
-
-      if (contextDescription != null) {
-        message.append(", ").append(contextDescription);
-      }
-
-      message.append(": ").append(issue);
-      throw new InvalidUserInputError(message.toString());
-    }
-
-    /**
-     * Checks a condition and reports an error if it's false.
-     *
-     * @param condition The condition to check
-     * @param issue The issue to report if the condition is false
-     * @throws InvalidUserInputError Thrown with a formatted message if condition is false
-     */
-    public void check(final boolean condition, @Nonnull final String issue) {
-      if (!condition) {
-        reportError(issue);
-      }
-    }
-  }
-
   /**
    * Creates a FunctionParameterResolver from a FunctionInput object.
    *
@@ -174,7 +55,6 @@ public class FunctionParameterResolver {
         functionInput.getInput(),
         functionInput.getArguments());
   }
-
 
   /**
    * Binds the input and arguments to the specified method.
