+++ fhirpath/src/test/java/au/csiro/pathling/fhirpath/function/FunctionParameterResolverTest.java
@@ -15,7 +15,8 @@ import au.csiro.pathling.fhirpath.collection.CodingCollection;
 import au.csiro.pathling.fhirpath.collection.Collection;
 import au.csiro.pathling.fhirpath.collection.IntegerCollection;
 import au.csiro.pathling.fhirpath.collection.StringCollection;
-import au.csiro.pathling.fhirpath.function.FunctionParameterResolver.FunctionInvocation;
+import au.csiro.pathling.fhirpath.function.resolver.FunctionInvocation;
+import au.csiro.pathling.fhirpath.function.resolver.FunctionParameterResolver;
 import au.csiro.pathling.fhirpath.path.ParserPaths;
 import jakarta.annotation.Nonnull;
 import jakarta.annotation.Nullable;
